+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- PublishedVersion did not handle publish\_to: none
+
 ## [3.0.4] - 2024-04-09
 
 ### Fixed
@@ -91,6 +97,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_publish/compare/3.0.4...HEAD
 [3.0.4]: https://github.com/inlavigo/gg_publish/compare/3.0.3...3.0.4
 [3.0.3]: https://github.com/inlavigo/gg_publish/compare/3.0.2...3.0.3
 [3.0.2]: https://github.com/inlavigo/gg_publish/compare/3.0.1...3.0.2